Simplify host account initalization
Initialize the assuming session before initializing the clients.

Inline each client initialization. A little repeated structure here is easier
to read than the complex type signatures of the deleted helper functions.

In a future change the STS client will take extra parameters to support version
2 STS tokens and opt-in regions.

This paves the way to a solution for #77.